Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.34.0
python-multipart==0.0.12
pdfplumber==0.11.4
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==24.1.0  
google-generativeai==0.1.0
openai==1.40.0